+++ src/lib/arch/win32/ArchMultithreadWindows.h
@@ -8,7 +8,7 @@
 #pragma once
 
 #include "arch/IArchMultithread.h"
-#include "common/stdlist.h"
+#include "common/StdList.h"
 
 #define WIN32_LEAN_AND_MEAN
 #include <Windows.h>
