+++ src/lib/io/StreamBuffer.h
@@ -8,7 +8,7 @@
 #pragma once
 
 #include "base/EventTypes.h"
-#include "common/stdlist.h"
+#include "common/StdList.h"
 #include "common/stdvector.h"
 
 //! FIFO of bytes
